requirements.txt
langchain
langgraph
langchain-groq
chromadb
sentence-transformers
yfinance
pandas
numpy
groq
pydantic
pytest
pandas-ta
matplotlib
python-dotenv